rename utils file, fix lint
